Update first_scene.md
While creating the scene for section 5 I've noticed some differences between the voiceover and the script. I think that script is wrong about the charge of the neutron. I decided to create this PR to fix this and keep things consistent with the voiceover. Please review the changes before merging.